requirements.txt
uvicorn
jsonpickle
joblib
fastapi
pymongo
python-dotenv